convnext-api/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
tensorflow
Pillow
python-multipart